pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stilwerk"
version = "0.2.0"
description = "Style analysis, attribution, and transformation toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "kai" }
]
keywords = ["style", "analysis", "stylometry", "authorship", "attribution", "transformation", "nlp", "writing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Text Processing :: Linguistic",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
]
viz = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
]

[project.scripts]
stilwerk = "stilwerk.src.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
